Agora cria screens dinamicos

+++ telas/passwd.py
@@ -11,6 +11,9 @@ from kivy.uix.button import Button
 from kivy.core.clipboard import Clipboard
 
 from kivy.uix.gridlayout import GridLayout
+
+
+from telas.login import Confirma
 #from telas.collect import JanelaEditCollect
 #from telas.collect import JanelaAddPass, JanelaEditCollect
 
@@ -38,10 +41,17 @@ class JanelaPassView (Screen):
             self.ids.pass_field.text = '******'
             self.ids.butt_show.text = 'show'
     
-    
+    def on_leave (self):
+        self.smanager.remove_widget(self)
+        
+    def _really_delete(self, really):
+        if really:
+            self.passwd.delete_instance(recursive=True)
+            self.voltar()
+            
     def delete (self):
-        self.passwd.delete_instance(recursive=True)
-        self.voltar()
+        p = Confirma (callback=self._really_delete)
+        p.open()
         
     def edit (self):
         janela = JanelaEditPass(smanager=self.smanager, name='janela_edit_pass') 
@@ -57,7 +67,10 @@ class JanelaPassView (Screen):
         self.ids.pass_field.text = ''
         self.ids.tx_desc.text = ''
         
-        janela = self.smanager.get_screen('janela_pass_list')
+        #janela = self.smanager.get_screen('janela_pass_list')
+        from telas.passwd import JanelaPassList
+        janela = JanelaPassList( smanager=self.smanager, name='janela_pass_list')
+        self.smanager.add_widget( janela )
         janela.setup(col=self.passwd.collect)
         self.smanager.transition.direction = 'right'
         self.smanager.current = 'janela_pass_list'
@@ -89,8 +102,15 @@ class JanelaPassList (Screen):
             b = ItemPass (passwd=s, smanager=self.smanager)
             self.ids.area_pass.add_widget(b)
     
+    def on_leave (self):
+        self.smanager.remove_widget(self)
+        
     def add (self):
-        janela = self.smanager.get_screen('janela_add_pass')
+        #janela = self.smanager.get_screen('janela_add_pass')
+        from telas.passwd import JanelaAddPass
+        janela = JanelaAddPass(smanager=self.smanager, name='janela_add_pass')
+        self.smanager.add_widget( janela )
+        
         janela.setup (col=self.collection)
         self.smanager.transition.direction = 'left'
         self.smanager.current = 'janela_add_pass'
@@ -104,10 +124,17 @@ class JanelaPassList (Screen):
         self.smanager.current = 'janela_edit_collect'
         
     def voltar (self):
-        janela = self.smanager.get_screen('janela_collect')
+        from telas.collect import JanelaCollect
+        janela = JanelaCollect(smanager=self.smanager, name='janela_collect')
+        self.smanager.add_widget( janela )
         janela.recarrega()
         self.smanager.transition.direction = 'right'
         self.smanager.current = 'janela_collect'
+        
+        #janela = self.smanager.get_screen('janela_collect')
+        #janela.recarrega()
+        #self.smanager.transition.direction = 'right'
+        #self.smanager.current = 'janela_collect'
     
     
 
@@ -122,8 +149,12 @@ class ItemPass (Button):
     def on_release (self, **kwargs):
         super(ItemPass, self).on_release(**kwargs)
         #jan = JanelaPassList( col=self.collection, name='janela_pass_list')
-        #self.manager.add_widget( jan )
-        janela = self.smanager.get_screen('janela_pass_view')
+        #self.smanager.add_widget( jan )
+        #janela = self.smanager.get_screen('janela_pass_view')
+        from telas.passwd import JanelaPassView
+        janela = JanelaPassView(smanager=self.smanager, name='janela_pass_view')
+        self.smanager.add_widget( janela )
+        
         janela.setup (passwd=self.passwd)
     
         self.smanager.transition.direction = 'left'
@@ -149,12 +180,19 @@ class JanelaAddPass (Screen):
         s.valor = self.smanager.encrypter.encripta (self.ids.tx_password.text)
         s.save()
         # Vai pra view
-        janela = self.smanager.get_screen('janela_pass_view')
+        from telas.passwd import JanelaPassView
+        janela = JanelaPassView(smanager=self.smanager, name='janela_pass_view')
+        self.smanager.add_widget( janela )
+        
+        #janela = self.smanager.get_screen('janela_pass_view')
         janela.setup (passwd=s)
         self.smanager.transition.direction = 'right'
         self.smanager.current = 'janela_pass_view'
        
        
+    def on_leave (self):
+        self.smanager.remove_widget(self)
+        
     def on_pre_enter(self):
         self.ids.tx_desc.text = ''
         self.ids.tx_password.text = ''
@@ -167,7 +205,10 @@ class JanelaAddPass (Screen):
         c.save()
         
     def voltar (self):
-        janela = self.smanager.get_screen('janela_pass_list')
+        #janela = self.smanager.get_screen('janela_pass_list')
+        from telas.passwd import JanelaPassList
+        janela = JanelaPassList(smanager=self.smanager, name='janela_pass_list')
+        self.smanager.add_widget( janela )
         janela.setup(col=self.collect)
         self.smanager.transition.direction = 'right'
         self.smanager.current = 'janela_pass_list'
@@ -189,7 +230,11 @@ class JanelaEditPass (JanelaAddPass):
         self.smanager.remove_widget (self)
         
     def voltar (self):
-        janela = self.smanager.get_screen('janela_pass_view')
+        #janela = self.smanager.get_screen('janela_pass_view')
+        from telas.passwd import JanelaPassView
+        janela = JanelaPassView(smanager=self.smanager, name='janela_pass_view')
+        self.smanager.add_widget( janela )
+        
         janela.setup (passwd=self.passwd)
         self.smanager.transition.direction = 'right'
         self.smanager.current = 'janela_pass_view'
@@ -202,7 +247,12 @@ class JanelaEditPass (JanelaAddPass):
             s.valor = self.smanager.encrypter.encripta (self.ids.tx_password.text)
         s.save()
         self.passwd = s
-        janela = self.smanager.get_screen('janela_pass_view')
+        #janela = self.smanager.get_screen('janela_pass_view')
+        
+        from telas.passwd import JanelaPassView
+        janela = JanelaPassView(smanager=self.smanager, name='janela_pass_view')
+        self.smanager.add_widget( janela )
+        
         janela.setup (passwd=s)
         self.smanager.transition.direction = 'right'
         self.smanager.current = 'janela_pass_view'
